[fix] use strict mode and named functions.
The strict mode allows the ansi parse to also be ran in environments that force strict
mode by default.

The named functions make it easier to debug and to be included in third party scripts
without polluting the global namespace.

In addition to this I have also made the === checks strict because I roll like that.